[feat] : 커뮤니티 바텀 시트 컨트롤러 설정 (#20)
