+++ lib/common/primary_app_bar.dart
@@ -1,8 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
+import 'package:get/get.dart';
 import 'package:mojacknong_android/common/farmus_theme_data.dart';
+import 'package:mojacknong_android/view_model/controllers/bottom_sheet_controller.dart';
 
 class PrimaryAppBar extends StatelessWidget implements PreferredSizeWidget {
+  final BottomSheetController _controller = Get.put(BottomSheetController());
+
   @override
   Size get preferredSize => Size.fromHeight(kToolbarHeight);
 
@@ -12,14 +16,24 @@ class PrimaryAppBar extends StatelessWidget implements PreferredSizeWidget {
       leading: IconButton(
         icon: SvgPicture.asset(
           "assets/image/ic_arrow_left.svg",
-          height: 24, // 원하는 크기로 조정
-          width: 24, // 원하는 크기로 조정
+          height: 24,
+          width: 24,
         ),
         onPressed: () {
-          Navigator.of(context).pop(); // 뒤로 가기 기능
+          Navigator.of(context).pop();
         },
       ),
-      actions: [],
+      actions: [
+        IconButton(
+          icon: SvgPicture.asset("assets/image/ic_more_vertical.svg"),
+          onPressed: () {
+            _controller.showCustomCupertinoActionSheet(context,
+                message: "글 메뉴",
+                options: ["수정", "삭제", "URL 공유"],
+                cancelButtonText: "취소");
+          },
+        ),
+      ],
       elevation: 0,
       backgroundColor: FarmusThemeData.white,
     );
